Updated display for a and b to 4 digit precision.
There is slight loss of precision as mentioned in issue #7 with some of
the Javascript math functions. This update forces a and b values to be
displayed to 4 digits of precision, to be consistent with the other
values on the page, and to mask the precision issues by rounding to 4
digits. This closes issue #7